Execute handles only one statement. Read updated to combine statements. Tests moved. Added error utility.
