+++ src/data/mssql/execute.js
@@ -7,8 +7,8 @@ var mssql = require('mssql'),
     log = require('../../logger'),
     connection, connectionPromise;
 
-module.exports = function (config, sql) {
-    if(sql.noop)
+module.exports = function (config, statement) {
+    if(statement.noop)
         return promises.resolved();
 
     if (!connectionPromise) {
@@ -25,29 +25,9 @@ module.exports = function (config, sql) {
     function executeRequest() {
         var request = new mssql.Request(connection);
 
-        var statements = sql;
-        if (sql.constructor !== Array) {
-            statements = [sql];
-        }
+        request.multiple = statement.multiple;
 
-        // We expect multiple results if there are multiple statements to be executed or if
-        // 'multiple' results are explicitly requested
-        request.multiple = statements.length > 1 ||  statements[0].multiple;
-
-        var statement = '',
-            params = [];
-
-        // Combine the statements into a single statement
-        statements.forEach(function(st) {
-            statement += st.sql + '; ';
-
-            if (st.parameters) {
-                params = params.concat(st.parameters)
-            }
-        });
-
-        // Combine the parameter lists into a single list
-        params.forEach(function (parameter) {
+        statement.parameters && statement.parameters.forEach(function (parameter) {
             var type = parameter.type || helpers.getMssqlType(parameter.value);
             if(type)
                 request.input(parameter.name, type, parameter.value);
@@ -55,10 +35,10 @@ module.exports = function (config, sql) {
                 request.input(parameter.name, parameter.value);
         });
 
-        log.silly('Executing SQL statement ' + statement + ' with parameters ' + JSON.stringify(params));
+        log.silly('Executing SQL statement ' + statement.sql + ' with parameters ' + JSON.stringify(statement.parameters));
 
-        return request.query(statement).catch(function (err) {
-            log.debug('SQL statement failed - ' + err.message + ': ' + statement + ' with parameters ' + JSON.stringify(params));
+        return request.query(statement.sql).catch(function (err) {
+            log.debug('SQL statement failed - ' + err.message + ': ' + statement.sql + ' with parameters ' + JSON.stringify(statement.parameters));
 
             if(err.number === 2627) {
                 var error = new Error('An item with the same ID already exists');
