+++ src/data/mssql/statements/read.js
@@ -1,15 +1,26 @@
 // ----------------------------------------------------------------------------
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // ----------------------------------------------------------------------------
-// translate the queryjs Query object into the odata format that our formatter expects
 var queries = require('../../../query'),
     format = require('../query/format');
 
 module.exports = function (source, tableConfig) {
+    // translate the queryjs Query object into the odata format that our formatter expects
     var query = queries.toOData(source);
 
     // copy custom properties from the source query. this is NOT ideal!
     query.includeDeleted = source.includeDeleted;
 
-    return format(query, tableConfig);
+    return combineStatements(format(query, tableConfig));
+};
+
+function combineStatements(statements) {
+    return statements.reduce(function(target, statement) {
+        target.sql += statement.sql + '; ';
+
+        if (statement.parameters)
+            target.parameters = target.parameters.concat(statement.parameters);
+
+        return target;
+    }, { sql: '', parameters: [], multiple: true });
 }
